requirements.txt
flask==3.0.3
pillow==10.4.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
gunicorn==22.0.0
